requirements.txt
Pillow==12.0.0
pygame-ce==2.5.6
